refactor(monte-carlo): split MonteCarloReport into focused files
Extract presentational helpers and scenario state/handlers out of the
2152-line MonteCarloReport so each file stays manageable:

- MonteCarloChartParts: CashFlowEvent type plus FanChartTooltip,
  CashFlowMarker, and CashFlowLegendSwatch.
- MonteCarloResultsTable: ResultsTable + SummaryStat.
- MonteCarloPerformanceSummary: SummaryRow type, header constants,
  buildPerformanceSummaryRows, formatSummaryValue, PerformanceSummaryTable.
- MonteCarloHoldingStatsTable: HoldingStatsTable.
- useMonteCarloScenarios: scenarios list, activeId, form, result,
  initial load + autofill effects, and all scenario CRUD handlers
  (load/new/run/save/saveAs/overwrite/delete + cancel handlers).

MonteCarloReport.tsx drops from 2152 to ~1260 lines and now only owns
chart-specific UI state, derived chart data, exports, and JSX.
